prepare-v5.0: TestData moved from SigmaDslSpecification.scala
