+++ sigmastate/src/test/scala/special/sigma/SigmaDslSpecification.scala
@@ -48,161 +48,6 @@ class SigmaDslSpecification extends SigmaDslTesting with CrossVersionProps { sui
 
   implicit def IR = createIR()
 
-  object TestData {
-    val BigIntZero: BigInt = CBigInt(new BigInteger("0", 16))
-    val BigIntOne: BigInt = CBigInt(new BigInteger("1", 16))
-    val BigIntMinusOne: BigInt = CBigInt(new BigInteger("-1", 16))
-    val BigInt10: BigInt = CBigInt(new BigInteger("a", 16))
-    val BigInt11: BigInt = CBigInt(new BigInteger("b", 16))
-
-    val ge1str = "03358d53f01276211f92d0aefbd278805121d4ff6eb534b777af1ee8abae5b2056"
-    val ge2str = "02dba7b94b111f3894e2f9120b577da595ec7d58d488485adf73bf4e153af63575"
-    val ge3str = "0290449814f5671172dd696a61b8aa49aaa4c87013f56165e27d49944e98bc414d"
-
-    val ge1 = Helpers.decodeGroupElement(ge1str)
-    val ge2 = Helpers.decodeGroupElement(ge2str)
-    val ge3 = Helpers.decodeGroupElement(ge3str)
-
-    val t1: AvlTree = CAvlTree(
-      AvlTreeData(
-        ADDigest @@ ErgoAlgos.decodeUnsafe("000183807f66b301530120ff7fc6bd6601ff01ff7f7d2bedbbffff00187fe89094"),
-        AvlTreeFlags(false, true, true),
-        1,
-        Some(1)
-      )
-    )
-    val t2: AvlTree = CAvlTree(
-      AvlTreeData(
-        ADDigest @@ ErgoAlgos.decodeUnsafe("ff000d937f80ffd731ed802d24358001ff8080ff71007f00ad37e0a7ae43fff95b"),
-        AvlTreeFlags(false, false, false),
-        32,
-        Some(64)
-      )
-    )
-    val t3: AvlTree = CAvlTree(
-      AvlTreeData(
-        ADDigest @@ ErgoAlgos.decodeUnsafe("3100d2e101ff01fc047c7f6f00ff80129df69a5090012f01ffca99f5bfff0c8036"),
-        AvlTreeFlags(true, false, false),
-        128,
-        None
-      )
-    )
-
-    val b1: Box = CostingBox(
-      false,
-      new ErgoBox(
-        9223372036854775807L,
-        new ErgoTree(
-          16.toByte,
-          Array(
-            SigmaPropConstant(
-              CSigmaProp(
-                ProveDlog(
-                  Helpers.decodeECPoint(
-                    "0297c44a12f4eb99a85d298fa3ba829b5b42b9f63798c980ece801cc663cc5fc9e"
-                  )
-                )
-              )
-            )
-          ),
-          Right(ConstantPlaceholder(0, SSigmaProp))
-        ),
-        Coll(
-          (Digest32 @@ (ErgoAlgos.decodeUnsafe("6e789ab7b2fffff12280a6cd01557f6fb22b7f80ff7aff8e1f7f15973d7f0001")),
-              10000000L),
-          (Digest32 @@ (ErgoAlgos.decodeUnsafe("a3ff007f00057600808001ff8f8000019000ffdb806fff7cc0b6015eb37fa600")),
-              500L)
-        ),
-        Map(
-          ErgoBox.R5 -> ByteArrayConstant(Helpers.decodeBytes("7fc87f7f01ff")),
-          ErgoBox.R4 -> FalseLeaf
-        ),
-        ModifierId @@ ("218301ae8000018008637f0021fb9e00018055486f0b514121016a00ff718080"),
-        22588.toShort,
-        677407
-      )
-    )
-
-    val b2: Box = CostingBox(
-      false,
-      new ErgoBox(
-        12345L,
-        new ErgoTree(
-          0.toByte,
-          Vector(),
-          Right(
-            BoolToSigmaProp(
-              AND(
-                ConcreteCollection(
-                  Array(
-                    FalseLeaf,
-                    XorOf(
-                      ConcreteCollection(Array(EQ(IntConstant(1), IntConstant(1)), FalseLeaf), SBoolean)
-                    )
-                  ),
-                  SBoolean
-                )
-              )
-            )
-          )
-        ),
-        Coll(),
-        Map(
-          ErgoBox.R5 -> ByteArrayConstant(
-            Helpers.decodeBytes(
-              "297000800b80f1d56c809a8c6affbed864b87f007f6f007f00ac00018c01c4fdff011088807f0100657f00f9ab0101ff6d65"
-            )
-          ),
-          ErgoBox.R4 -> TrueLeaf,
-          ErgoBox.R7 -> LongConstant(9223372036854775807L),
-          ErgoBox.R6 -> LongConstant(2115927197107005906L)
-        ),
-        ModifierId @@ ("003bd5c630803cfff6c1ff7f7fb980ff136afc011f8080b8b04ad4dbda2d7f4e"),
-        1.toShort,
-        1000000
-      )
-    )
-
-    val preH1: PreHeader = CPreHeader(
-      0.toByte,
-      Helpers.decodeBytes("7fff7fdd6f62018bae0001006d9ca888ff7f56ff8006573700a167f17f2c9f40"),
-      6306290372572472443L,
-      -3683306095029417063L,
-      1,
-      Helpers.decodeGroupElement("026930cb9972e01534918a6f6d6b8e35bc398f57140d13eb3623ea31fbd069939b"),
-      Helpers.decodeBytes("ff8087")
-    )
-
-    val preH2: PreHeader = preH1.asInstanceOf[CPreHeader].copy(height = 2)
-
-    val treeData = AvlTreeData(
-      ADDigest @@ (
-          ErgoAlgos.decodeUnsafe("010180017f7f7b7f720c00007f7f7f0f01e857a626f37f1483d06af8077a008080")
-          ),
-      AvlTreeFlags(false, true, false),
-      728138553,
-      Some(2147483647)
-    )
-    val h1: Header = CHeader(
-      Helpers.decodeBytes("957f008001808080ffe4ffffc8f3802401df40006aa05e017fa8d3f6004c804a"),
-      0.toByte,
-      Helpers.decodeBytes("0180dd805b0000ff5400b997fd7f0b9b00de00fb03c47e37806a8186b94f07ff"),
-      Helpers.decodeBytes("01f07f60d100ffb970c3007f60ff7f24d4070bb8fffa7fca7f34c10001ffe39d"),
-      CAvlTree(treeData),
-      Helpers.decodeBytes("804101ff01000080a3ffbd006ac080098df132a7017f00649311ec0e00000100"),
-      1L,
-      -1L,
-      1,
-      Helpers.decodeBytes("e57f80885601b8ff348e01808000bcfc767f2dd37f0d01015030ec018080bc62"),
-      Helpers.decodeGroupElement("039bdbfa0b49cc6bef58297a85feff45f7bbeb500a9d2283004c74fcedd4bd2904"),
-      Helpers.decodeGroupElement("0361299207fa392231e23666f6945ae3e867b978e021d8d702872bde454e9abe9c"),
-      Helpers.decodeBytes("7f4f09012a807f01"),
-      CBigInt(new BigInteger("-e24990c47e15ed4d0178c44f1790cc72155d516c43c3e8684e75db3800a288", 16)),
-      Helpers.decodeBytes("7f0180")
-    )
-    val h2: Header = h1.asInstanceOf[CHeader].copy(height = 2)
-
-  }
   import TestData._
 
   prepareSamples[BigInt]
@@ -3581,7 +3426,7 @@ class SigmaDslSpecification extends SigmaDslTesting with CrossVersionProps { sui
       existingPropTest("ADProofsRoot", { (x: Header) => x.ADProofsRoot}))
 
     verifyCases(
-      Seq((h1, Expected(Success(CAvlTree(treeData)), cost = 36092))),
+      Seq((h1, Expected(Success(CAvlTree(createAvlTreeData())), cost = 36092))),
       existingPropTest("stateRoot", { (x: Header) => x.stateRoot }))
 
     verifyCases(
